+++ src/main/java/duke/ui/MainWindow.java
@@ -1,6 +1,6 @@
-package duke;
+package duke.ui;
 
-import duke.ui.Ui;
+import duke.Duke;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
